big optimization of the orders and products

+++ src/app/Http/Controllers/SupplierController.php
@@ -5,72 +5,86 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Requests\SupplierRequest;
-use App\Services\SupplierService;
 use App\Models\Supplier;
 use App\Models\SupplierCategory;
 use App\Helpers\LoadStaticData;
+use App\Models\SupplierImage;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Storage;
 
-class SupplierController extends Controller {
+class SupplierController extends Controller
+{
 
     private $staticDataHelper;
+    private $dir = 'public/images/suppliers';
 
     public function __construct(LoadStaticData $staticDataHelper)
     {
         $this->staticDataHelper = $staticDataHelper;
     }
 
-    public function index() {
+    public function index()
+    {
         $countries = $this->staticDataHelper->callStatesAndCountries('countries');
         $states = $this->staticDataHelper->callStatesAndCountries('states');
         $categories = $this->staticDataHelper->loadCallCategories();
-        
-        return view('suppliers.index',[
+
+        return view('suppliers.index', [
             'countries' => $countries,
             'states' => $states,
             'categories' => $categories
         ]);
     }
 
-    public function create() {
+    public function create()
+    {
         $countries = $this->staticDataHelper->callStatesAndCountries("countries");
         $categories = $this->staticDataHelper->loadCallCategories();
 
-        return view('suppliers.create', ["countries" => $countries,"categories"=>$categories]);
+        return view('suppliers.create', ["countries" => $countries, "categories" => $categories]);
     }
 
-    public function getState($countryId) {
+    public function getState($countryId)
+    {
         return response()->json(
-            $this->staticDataHelper->callStatesAndCountries($countryId,'states')
+            $this->staticDataHelper->callStatesAndCountries($countryId, 'states')
         );
     }
 
-    public function store(SupplierRequest $request) {
+    public function store(SupplierRequest $request)
+    {
         $data = $request->validated();
-        
+
         DB::beginTransaction();
- 
+
         try {
+            $file = isset($data['image']) ? $data['image'] : false;
+            $categories = isset($data['categories']) ? $data['categories'] : false ;
+            $imagePath = Storage::url($this->dir);
+
             $supplier = Supplier::create($data);
 
-            if ($supplier) {
-                $supplierService = new SupplierService($supplier);
+            if (isset($categories) && !empty($categories)) {
+                $supplier->categories()->sync($categories);
+            }
 
-                if($request->file('image')){
-                    $supplierService->imageUploader($request->file('image'));
-                }
-                            
-                if ($supplier && isset($data['categories']) && count($data['categories'])) {
-                    $supplierService->attachSupplierCategories( $data['categories'] );
-                }
+            if ($file) {
+                $hashed_image = Str::random(10).'.'.$file->getClientOriginalExtension();
+                Storage::putFileAs($this->dir, $file, $hashed_image);
+
+                $image = new SupplierImage([
+                    'path' => $imagePath,
+                    'name' => $hashed_image
+                ]);
+
+                $supplier->image()->save($image);
             }
 
             DB::commit();
-
-            Log::info('Successfully created supplier');
         } catch (\Exception $e) {
             DB::rollback();
-            Log::error($e->getMessage());
+            return response()->json(['message' => 'Supplier has not been created'], 500);
         }
 
         return redirect()->route('supplier.index')->with('success', 'Supplier has been created');
@@ -78,11 +92,10 @@ class SupplierController extends Controller {
 
     public function edit(Supplier $supplier)
     {
-        $service = new SupplierService($supplier);
         $country = $supplier->country_id;
 
         $categories = $this->staticDataHelper->loadCallCategories();
-        $states = $this->staticDataHelper->callStatesAndCountries($country,'states');
+        $states = $this->staticDataHelper->callStatesAndCountries($country, 'states');
         $countries = $this->staticDataHelper->callStatesAndCountries('countries');
 
         $supplier = $supplier->load('image');
@@ -95,39 +108,53 @@ class SupplierController extends Controller {
         ]);
     }
 
-    public function update(Supplier $supplier, SupplierRequest $request) {
-
-        $supplierService = new SupplierService($supplier);
+    public function update(Supplier $supplier, SupplierRequest $request)
+    {
+        $data = $request->validated();
 
         DB::beginTransaction();
 
         try {
-            
-            if ($request->hasFile('image')) {
-                $supplierService->imageUploader($request->file('image'));
+            $file = isset($data['image']) ? $data['image'] : false;
+            $categories = isset($data['categories']) ? $data['categories'] : false ;
+            $imagePath = Storage::url($this->dir);
+
+            if (isset($categories) && !empty($categories)) {
+                $supplier->categories()->sync($categories);
             }
-                        
-            if (isset($request->categories) && count($request->categories)) {
-                $supplierService->attachSupplierCategories($request->categories);
+
+            if($file) {
+                $hashed_image = Str::random(10).'.'.$file->getClientOriginalExtension();
+                $current_image = null;
+
+                if($supplier->image) {
+                    $current_image = $this->dir.DIRECTORY_SEPARATOR.$supplier->image->name;
+                 
+                    if(Storage::exists($current_image)) {
+                        Storage::delete($current_image);
+                    } 
+
+                    Storage::putFileAs($this->dir, $file, $hashed_image);
+
+                    $supplier->image->name = $hashed_image;
+                    $supplier->image->path = $imagePath;
+                    $supplier->image->save();
+                } else {
+                    Storage::putFileAs($this->dir, $file, $hashed_image);
+                    
+                    $image = new SupplierImage([
+                        'path' => $imagePath,
+                        'name' => $hashed_image
+                    ]);
+
+                    $supplier->image()->save($image);
+                }            
             }
-            
-            $supplier->update([
-                'name' => $request->name,
-                'email' => $request->email,
-                'phone' => $request->phone,
-                'address' => $request->address,
-                'zip' => $request->zip,
-                'website' => $request->website,
-                'state_id' => $request->state_id,
-                'country_id' => $request->country_id,
-                'notes' => $request->notes,
-            ]);            
+
+            $supplier->update($data);
 
             DB::commit();
-
-            Log::info('Succesfully updated supplier');
         } catch (\Exception $e) {
-            dd($e->getMessage());
             DB::rollback();
             Log::error($e->getMessage());
         }
@@ -136,32 +163,36 @@ class SupplierController extends Controller {
         return redirect()->route('supplier.index')->with('success', 'Supplier has been updated');
     }
 
-    public function delete(Supplier $supplier) {
-        
+    public function delete(Supplier $supplier)
+    {
         DB::beginTransaction();
         try {
-            $supplierImage = $supplier->image;
 
-            if ($supplierImage) {
-                $imagePath = storage_path('app/public/images/suppliers/' . $supplierImage->name);
+            if ($supplier->purchases()->exists()) {
+                return response()->json(['message' => 'Supplier has related products'], 500);
+            }
+
+            if($supplier->image) {
+                $current_image = $this->dir.DIRECTORY_SEPARATOR.$supplier->image->name;
                 
-                if (file_exists($imagePath)) {
-                    unlink($imagePath);
-                }
-
-            }   
-                     
+                if(Storage::exists($current_image)) {
+                    Storage::delete($current_image);
+                };
+            }
+            
             $supplier->delete();
             DB::commit();
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => 'Supplier has not been deleted'],500);
+            Log::error($e->getMessage());
+            return response()->json(['message' => 'Supplier has not been deleted'], 500);
         }
 
         return response()->json(['message' => 'Supplier has been deleted'], 200);
     }
-    
-    public function detachCategory($id) {
+
+    public function detachCategory($id)
+    {
         DB::beginTransaction();
 
         try {
@@ -169,10 +200,15 @@ class SupplierController extends Controller {
             $related_category->delete();
             DB::commit();
         } catch (\Exception $e) {
-            DB::rollback(); 
+            DB::rollback();
             return response()->json(['message' => 'Category has not been detached'], 500);
         }
         return response()->json(['message' => 'Category has been detached'], 200);
     }
 
+    public function massEdit(Supplier $supplier){
+        $categories = $supplier->categories()->get();
+        $brands = $this->staticDataHelper->callBrands();
+        return view('suppliers.mass_edit_purchases',['supplier' => $supplier,'categories' => $categories,'brands' => $brands]);
+    }
 }
